+++ crypto/src/jvmTest/kotlin/at/asitplus/awesn1/crypto/RealWorldCertificateTest.kt
@@ -7,8 +7,8 @@ import at.asitplus.awesn1.decodeAllFromPem
 import at.asitplus.awesn1.serialization.DER
 import at.asitplus.testballoon.withData
 import de.infix.testBalloon.framework.core.testSuite
-import io.kotest.matchers.collections.shouldContain
 import io.kotest.matchers.collections.shouldContainAll
+import java.io.ByteArrayInputStream
 import java.nio.file.Files
 import java.nio.file.Path
 import kotlin.io.path.extension
@@ -16,10 +16,11 @@ import kotlin.io.path.nameWithoutExtension
 import kotlin.io.path.readText
 
 val RealWorldCertificateTest by testSuite {
-    val root = object{}.javaClass.classLoader.getResource("real-world-certs")?.toURI()?.let(Path::of)
+    val root = object {}.javaClass.classLoader.getResource("real-world-certs")?.toURI()?.let(Path::of)
         ?: throw IllegalStateException("Missing real-world-certs dir in resources")
 
-    withData(nameFn = Path::nameWithoutExtension,
+    withData(
+        nameFn = Path::nameWithoutExtension,
         Files.walk(root)
             .filter(Files::isRegularFile)
             .filter { it.extension == "pem" }
@@ -33,6 +34,12 @@ val RealWorldCertificateTest by testSuite {
             .first { it.label == X509Certificate.PEM_LABEL }
             .payload
             .let { DER.decodeFromByteArray(X509Certificate.serializer(), it) }
+        val jvmCert = certificateFactory.generateCertificate(
+            ByteArrayInputStream(
+            PemBlock.decodeAllFromPem(certPEM)
+            .first { it.label == X509Certificate.PEM_LABEL }
+            .payload))
+        assertEquals(cert, jvmCert as java.security.cert.X509Certificate)
         cert.findSubjectAltNames()!!.dnsNames shouldContainAll listOf(tld, "*.$tld")
     }
 }